% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment_to_ref.R
\name{alignment_to_ref}
\alias{alignment_to_ref}
\title{Align read to reference genome based on CIGAR string}
\usage{
alignment_to_ref(read, genome = NULL)
}
\arguments{
\item{read}{A specific read extracted from a BAM file.}

\item{genome}{An optional reference genome given in input by the user.}
}
\value{
The function return a list object reporting the read ID, CIGAR string
and the alignment of the query read to the reference genome.
}
\description{
This function takes as input a specific read extracted from a BAM file and
a reference genome, if no reference genome is given in input,
a UCSC hg38 reference genome is used, and return the alignment of the read
to the reference genome, using the CIGAR of the specific read.
All the mapped reads, including those that have been mapped to the
reverse strand,in the BAM file are represented on the forward
genomic strand and also CIGAR string is reversed and thus recorded
consistently with the sequence bases as represented. Taking this into
consideration the alignment of a query read, that have been mapped to the
reverse strand, to the genome is also done on the forward genomic strand.
}
\examples{
example_read <- list(rname = "chr6",
pos = 32402000,
strand = "+",
cigar = "2S5M2D4M",
seq = "AAAGATCGACC",
qname = "read4")
alignment_to_ref(example_read)
# Returns: $Read_ID
# [1] "read4"
#
# $CIGAR
# [1] "2S5M2D4M"
#
# $Alignment
# $Alignment$Reference_genome
# [1] "  AGATCGAGACC"
#
# $Alignment$Query_read
# [1] "AAAGATC--GACC"
}
