% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cigar_transform.R
\name{cigar_transform}
\alias{cigar_transform}
\title{Transform CIGAR into a character vector}
\usage{
cigar_transform(cigar)
}
\arguments{
\item{cigar}{A CIGAR string extracted from a BAM file.}
}
\value{
A character vector where each element represents an individual
operation from the CIGAR string.
}
\description{
This function takes a CIGAR string as input and transforms
it into a character vector
by expanding each operation symbol according to the number of times
it occurs.
}
\examples{
cigar_transform("2M1I3M1D")
# Returns: [1] "M" "M" "I" "M" "M" "M" "D"
}
